+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2 
+- Changed constructor of binary expression tree
+- Changed constructor of node
+- Added more tests
+- Remove unused imports
+
 ## 1.0.1
 
 - Refactor and removal of unnecessary code